Create Homework5
A Personal Fitness Tracker is a wearable device that tracks your physical activity, calories burned, heart rate, sleeping patterns, and so on. One common physical activity that most of these devices track is the number of steps you take each day.

On the course moodle page is a file called steps.txt
Save this file as ‘steps.txt’ in your working python directory.
This steps.txt file has the following format:
MMM,#Steps   where
	- MMM – is a 3 digit month code (e.g., Jan, Feb, Mar, Apr, May,  Jun, Jul, Aug, Sep, Oct, Nov, Dec)
	- #Steps – Is the number of steps that person took in a day of that month

Write a program that reads the file, then displays the average number of steps taken for each month and the overall average of steps per day.